+++ nuls-switch-backend/src/main/resources/db/nuls_switch_init.sql
@@ -77,7 +77,7 @@ CREATE TABLE `tx_trade` (
   `tx_id` varchar(255) NOT NULL COMMENT '交易流水号',
   `order_id` varchar(32) NOT NULL COMMENT '委托挂单ID',
   `address` varchar(255) DEFAULT NULL COMMENT '交易用户地址',
-  `tx_num` int(11) DEFAULT NULL COMMENT '交易数量',
+  `tx_num` bigint(20) DEFAULT NULL COMMENT '交易数量',
   `tx_hash` varchar (255) DEFAULT NULL COMMENT '交易发送到区块链后的Hash值',
   `status` tinyint(4) NOT NULL COMMENT '状态：0-未交易、1-完成交易、9-撤销',
   `create_time` datetime DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
